requirements.txt
beaker
sanic
requests
kafka-python
influxdb
Geohash
geoip2
python-geohash